Extract Prompt Studio DOM helpers
